Move storage adapter and traced ES client into dedicated package (#211578)
The storage adapter helper is a very generic package. This PR moves it
out of the observability server utils into a dedicated package to better
reflect this and to be able to use it from non-observability contexts.

The same applies to the observability es client. This PR moves it as
well and renames it to `TracedEsClient` in the same way.

---------

(cherry picked from commit cb9c5f21b2dbdb2a03c71767abe45b5ad7f2ed6b)

# Conflicts:
#	.github/CODEOWNERS
#	x-pack/solutions/observability/plugins/infra/tsconfig.json
#	x-pack/solutions/observability/plugins/investigate_app/tsconfig.json
#	x-pack/solutions/observability/plugins/observability_ai_assistant_app/tsconfig.json
#	x-pack/solutions/observability/plugins/streams_app/tsconfig.json
