+++ .github/CODEOWNERS
@@ -955,6 +955,7 @@ test/server_integration/plugins/status_plugin_a @elastic/kibana-core
 test/server_integration/plugins/status_plugin_b @elastic/kibana-core
 src/platform/packages/shared/kbn-std @elastic/kibana-core
 packages/kbn-stdio-dev-helpers @elastic/kibana-operations
+src/platform/packages/shared/kbn-storage-adapter @elastic/observability-ui
 packages/kbn-storybook @elastic/kibana-operations
 x-pack/solutions/observability/plugins/streams_app @simianhacker @flash1293 @dgieselaar
 x-pack/solutions/observability/plugins/streams @simianhacker @flash1293 @dgieselaar
@@ -990,6 +991,7 @@ src/platform/packages/private/kbn-timelion-grammar @elastic/kibana-visualization
 src/platform/packages/shared/kbn-timerange @elastic/obs-ux-logs-team
 src/platform/packages/private/kbn-tinymath @elastic/kibana-visualizations
 packages/kbn-tooling-log @elastic/kibana-operations
+src/platform/packages/shared/kbn-traced-es-client @elastic/observability-ui
 x-pack/platform/plugins/private/transform @elastic/ml-ui
 x-pack/platform/plugins/private/translations @elastic/kibana-localization
 src/platform/packages/private/kbn-transpose-utils @elastic/kibana-visualizations
