+++ src/platform/packages/shared/kbn-storage-adapter/index.ts
@@ -1,9 +1,12 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
  */
+
 import type {
   BulkRequest,
   BulkResponse,
@@ -151,6 +154,6 @@ export type StorageDocumentOf<TStorageSettings extends StorageSettings> = Storag
   properties: TStorageSettings['schema']['properties'];
 }> & { _id: string };
 
-export { StorageIndexAdapter } from './index_adapter';
+export { StorageIndexAdapter } from './src/index_adapter';
 
 export { types } from './types';
