+++ src/platform/packages/shared/kbn-storage-adapter/src/get_schema_version.ts
@@ -1,0 +1,17 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import stringify from 'json-stable-stringify';
+import objectHash from 'object-hash';
+import { IndexStorageSettings } from '..';
+
+export function getSchemaVersion(storage: IndexStorageSettings): string {
+  const version = objectHash(stringify(storage.schema.properties));
+  return version;
+}
