+++ src/platform/packages/shared/kbn-storage-adapter/src/index_adapter/index.ts
@@ -1,8 +1,10 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
 import type {
@@ -35,9 +37,9 @@ import {
   StorageClientCleanResponse,
   ApplicationDocument,
   InternalIStorageClient,
-} from '..';
+} from '../..';
 import { getSchemaVersion } from '../get_schema_version';
-import { StorageMappingProperty } from '../types';
+import { StorageMappingProperty } from '../../types';
 
 function getAliasName(name: string) {
   return name;
