+++ src/platform/packages/shared/kbn-traced-es-client/README.md
@@ -1,0 +1,4 @@
+# Traced ES Client
+
+This package provides a wrapper around the regular platform ES client which adds tracing information for better analysis.
+Each operation can be given a custom name that shows up as a trace when instrumenting using APM.
