+++ src/platform/packages/shared/kbn-traced-es-client/kibana.jsonc
@@ -1,0 +1,9 @@
+{
+  "type": "shared-server",
+  "id": "@kbn/traced-es-client",
+  "owner": [
+    "@elastic/observability-ui"
+  ],
+  "group": "platform",
+  "visibility": "shared"
+}
