+++ src/platform/packages/shared/kbn-traced-es-client/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@kbn/traced-es-client",
+  "private": true,
+  "version": "1.0.0",
+  "license": "Elastic License 2.0 OR AGPL-3.0-only OR SSPL-1.0"
+}
