+++ src/platform/packages/shared/kbn-traced-es-client/tsconfig.json
@@ -16,14 +16,10 @@
     "target/**/*"
   ],
   "kbn_references": [
+    "@kbn/core",
     "@kbn/es-types",
-    "@kbn/es-query",
-    "@kbn/observability-utils-common",
-    "@kbn/traced-es-client",
-    "@kbn/alerting-plugin",
-    "@kbn/rule-registry-plugin",
-    "@kbn/rule-data-utils",
-    "@kbn/aiops-log-pattern-analysis",
-    "@kbn/calculate-auto",
+    "@kbn/apm-utils",
+    "@kbn/utility-types",
+    "@kbn/task-manager-plugin",
   ]
 }
