+++ tsconfig.base.json
@@ -1904,6 +1904,8 @@
       "@kbn/std/*": ["src/platform/packages/shared/kbn-std/*"],
       "@kbn/stdio-dev-helpers": ["packages/kbn-stdio-dev-helpers"],
       "@kbn/stdio-dev-helpers/*": ["packages/kbn-stdio-dev-helpers/*"],
+      "@kbn/storage-adapter": ["src/platform/packages/shared/kbn-storage-adapter"],
+      "@kbn/storage-adapter/*": ["src/platform/packages/shared/kbn-storage-adapter/*"],
       "@kbn/storybook": ["packages/kbn-storybook"],
       "@kbn/storybook/*": ["packages/kbn-storybook/*"],
       "@kbn/streams-app-plugin": ["x-pack/solutions/observability/plugins/streams_app"],
@@ -1974,6 +1976,8 @@
       "@kbn/tinymath/*": ["src/platform/packages/private/kbn-tinymath/*"],
       "@kbn/tooling-log": ["packages/kbn-tooling-log"],
       "@kbn/tooling-log/*": ["packages/kbn-tooling-log/*"],
+      "@kbn/traced-es-client": ["src/platform/packages/shared/kbn-traced-es-client"],
+      "@kbn/traced-es-client/*": ["src/platform/packages/shared/kbn-traced-es-client/*"],
       "@kbn/transform-plugin": ["x-pack/platform/plugins/private/transform"],
       "@kbn/transform-plugin/*": ["x-pack/platform/plugins/private/transform/*"],
       "@kbn/translations-plugin": ["x-pack/platform/plugins/private/translations"],
