+++ x-pack/solutions/observability/packages/observability-ai/observability-ai-server/root_cause_analysis/run_root_cause_analysis.ts
@@ -15,7 +15,7 @@ import { findLast, pick } from 'lodash';
 import moment from 'moment';
 import { catchError, filter, from, map, mergeMap, Observable, of, switchMap } from 'rxjs';
 import { ObservabilityAIAssistantClient } from '@kbn/observability-ai-assistant-plugin/server';
-import { ObservabilityElasticsearchClient } from '@kbn/observability-utils-server/es/client/create_observability_es_client';
+import { TracedElasticsearchClient } from '@kbn/traced-es-client';
 import {
   RCA_END_PROCESS_TOOL_NAME,
   RCA_INVESTIGATE_ENTITY_TOOL_NAME,
@@ -94,7 +94,7 @@ export function runRootCauseAnalysis({
   end: number;
   alertsClient: AlertsClient;
   rulesClient: RulesClient;
-  esClient: ObservabilityElasticsearchClient;
+  esClient: TracedElasticsearchClient;
   observabilityAIAssistantClient: ObservabilityAIAssistantClient;
   indices: {
     logs: string[];
