+++ x-pack/solutions/observability/packages/observability-ai/observability-ai-server/tsconfig.json
@@ -25,5 +25,6 @@
     "@kbn/inference-common",
     "@kbn/observability-ai-common",
     "@kbn/observability-utils-server",
+    "@kbn/traced-es-client",
   ]
 }
