+++ x-pack/solutions/observability/packages/utils-server/entities/signals/get_slos_for_entity.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { ObservabilityElasticsearchClient } from '../../es/client/create_observability_es_client';
+import { TracedElasticsearchClient } from '@kbn/traced-es-client';
 import { kqlQuery } from '../../es/queries/kql_query';
 
 export async function getSlosForEntity({
@@ -20,7 +20,7 @@ export async function getSlosForEntity({
   start: number;
   end: number;
   entity: Record<string, unknown>;
-  esClient: ObservabilityElasticsearchClient;
+  esClient: TracedElasticsearchClient;
   sloSummaryIndices: string | string[];
   size: number;
   spaceId: string;
