+++ x-pack/solutions/observability/plugins/infra/server/routes/entities/get_data_stream_types.ts
@@ -9,7 +9,7 @@ import { type EntityClient } from '@kbn/entityManager-plugin/server/lib/entity_c
 import type { InventoryItemType } from '@kbn/metrics-data-access-plugin/common';
 import { findInventoryFields } from '@kbn/metrics-data-access-plugin/common';
 import { EntityDataStreamType } from '@kbn/observability-shared-plugin/common';
-import type { ObservabilityElasticsearchClient } from '@kbn/observability-utils-server/es/client/create_observability_es_client';
+import type { TracedElasticsearchClient } from '@kbn/traced-es-client';
 import { castArray } from 'lodash';
 import type { Logger } from '@kbn/logging';
 import { type InfraMetricsClient } from '../../lib/helpers/get_infra_metrics_client';
@@ -22,7 +22,7 @@ interface Params {
   entityFilterType: string;
   entityCentricExperienceEnabled: boolean;
   infraMetricsClient: InfraMetricsClient;
-  obsEsClient: ObservabilityElasticsearchClient;
+  obsEsClient: TracedElasticsearchClient;
   entityManagerClient: EntityClient;
   logger: Logger;
   from: string;
