+++ x-pack/solutions/observability/plugins/infra/tsconfig.json
@@ -115,12 +115,13 @@
     "@kbn/core-ui-settings-common",
     "@kbn/entityManager-plugin",
     "@kbn/zod",
-    "@kbn/observability-utils-server",
-    "@kbn/observability-utils-common",
     "@kbn/core-plugins-server",
     "@kbn/config",
     "@kbn/response-ops-rule-params",
-    "@kbn/charts-theme"
+    "@kbn/charts-theme",
+    "@kbn/core-test-helpers-model-versions",
+    "@kbn/deeplinks-management",
+    "@kbn/traced-es-client"
   ],
   "exclude": ["target/**/*"]
 }
