+++ x-pack/solutions/observability/plugins/observability_ai_assistant_app/server/clients/create_observability_ai_assistant_app_es_client.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { createObservabilityEsClient } from '@kbn/observability-utils-server/es/client/create_observability_es_client';
+import { createTracedEsClient } from '@kbn/traced-es-client';
 import { ElasticsearchClient, Logger } from '@kbn/core/server';
 
 export async function createObservabilityAIAssistantAppEsClient({
@@ -15,7 +15,7 @@ export async function createObservabilityAIAssistantAppEsClient({
   client: ElasticsearchClient;
   logger: Logger;
 }) {
-  const esClient = createObservabilityEsClient({
+  const esClient = createTracedEsClient({
     client,
     logger,
     plugin: 'observabilityAIAssistantApp',
