+++ x-pack/solutions/observability/plugins/streams/server/lib/streams/assets/storage_settings.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { IndexStorageSettings, types } from '@kbn/observability-utils-server/es/storage';
+import { IndexStorageSettings, types } from '@kbn/storage-adapter';
 import { ASSET_ASSET_ID, ASSET_ENTITY_ID, ASSET_ENTITY_TYPE, ASSET_TYPE } from './fields';
 
 export const assetStorageSettings = {
