+++ x-pack/solutions/observability/plugins/streams/server/routes/esql/route.ts
@@ -8,10 +8,7 @@
 import { excludeFrozenQuery } from '@kbn/observability-utils-common/es/queries/exclude_frozen_query';
 import { kqlQuery } from '@kbn/observability-utils-common/es/queries/kql_query';
 import { rangeQuery } from '@kbn/observability-utils-common/es/queries/range_query';
-import {
-  UnparsedEsqlResponse,
-  createObservabilityEsClient,
-} from '@kbn/observability-utils-server/es/client/create_observability_es_client';
+import { UnparsedEsqlResponse, createTracedEsClient } from '@kbn/traced-es-client';
 import { z } from '@kbn/zod';
 import { isNumber } from 'lodash';
 import { createServerRoute } from '../create_server_route';
@@ -40,7 +37,7 @@ export const executeEsqlRoute = createServerRoute({
   }),
   handler: async ({ params, request, logger, getScopedClients }): Promise<UnparsedEsqlResponse> => {
     const { scopedClusterClient } = await getScopedClients({ request });
-    const observabilityEsClient = createObservabilityEsClient({
+    const observabilityEsClient = createTracedEsClient({
       client: scopedClusterClient.asCurrentUser,
       logger,
       plugin: 'streams',
