+++ x-pack/solutions/observability/plugins/streams/tsconfig.json
@@ -36,6 +36,8 @@
     "@kbn/safer-lodash-set",
     "@kbn/streams-schema",
     "@kbn/es-errors",
-    "@kbn/server-route-repository-utils"
+    "@kbn/server-route-repository-utils",
+    "@kbn/storage-adapter",
+    "@kbn/traced-es-client"
   ]
 }
